refactor: rename scorer and file name
